+++ rinla/R/collect.results.R
@@ -2273,7 +2273,7 @@
     ## reduce image IM to image.dim IMAGE.DIM and return the image as a matrix.
     ## order the indices so the output can be plotted by image()
 
-    if ((class(im) != "pixmapGrey") || (im@size[1L] != im@size[2L])) {
+    if (!inherits(im, "pixmapGrey") || (im@size[1L] != im@size[2L])) {
         return(im)
     } else {
         return(im@grey)